♻️ Move story files to appropriate hierarchy
