+++ client/src/registry.ts
@@ -2,7 +2,7 @@ import MobileLayout from '@layouts/mobile'
 import Button from '@elements/button'
 import Canvas from '@elements/canvas'
 import Icon from '@elements/icon'
-import IconButton from '@elements/icon-button'
+import IconButton from '@/components/molecules/icon-button'
 import Container from '@elements/container'
 import TextInput from '@elements/text-input'
 import ControlContainer from '@molecules/control-container'
