+++ client/src/stories/molecules/icon-button.stories.ts
@@ -1,7 +1,7 @@
 import { html } from 'lit-html'
 
-import IconButton from '@elements/icon-button'
-import type { Icon, Size } from '@elements/icon-button'
+import IconButton from '@/components/molecules/icon-button'
+import type { Icon, Size } from '@/components/molecules/icon-button'
 
 customElements.define(IconButton.tag, IconButton)
